+++ Disciple/src/utopia/disciple/http/request/FileBody.scala
@@ -1,10 +1,8 @@
 package utopia.disciple.http.request
 
-import java.io.{File, FileInputStream}
-import java.nio.charset.Charset
-
 import utopia.access.http.ContentType
 
+import java.io.{File, FileInputStream}
 import scala.util.Try
 
 /**
@@ -12,8 +10,7 @@ import scala.util.Try
 * @author Mikko Hilpinen
 * @since 1.5.2018
 **/
-class FileBody(val file: File, val contentType: ContentType, val charset: Option[Charset] = None, 
-        val contentEncoding: Option[String] = None) extends Body
+class FileBody(val file: File, val contentType: ContentType, val contentEncoding: Option[String] = None) extends Body
 {
 	override def repeatable = true
 	override def chunked = false