+++ Scribe/Scribe-Client/Changes.md
@@ -1,6 +1,7 @@
 # Utopia Scribe Client - list of changes
 
 ## v1.0.2 (in development)
+Supports Flow & Annex updates
 ### Other changes
 - In some instances where **Vector** was used, **Seq** is now used